Removed ability to declare metrics to get
Determined the format of the metrics file is too difficult to predict how each metric may be formatted. This causes a lot of trouble with trying to recover specific metrics. It would be easier to have a static algorithm to extract all metrics available in the metrics file and maintain this with future versions of Integrity.